separated chat and video
