+++ src/api/tools.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/prefer-default-export
 export const getTimeStamp = () => {
 	const date = new Date();
-	return `${date.getHours()}:${date.getMinutes()}:${date.getSeconds()}`;
+	return `${date.getHours()}:${date.getMinutes()}`;
 };
