+++ src/components/Chat/Chat.jsx
@@ -2,16 +2,13 @@ import React, { useEffect, useState } from 'react';
 import Peer from 'simple-peer';
 import { getTimeStamp } from '../../api/tools';
 import styles from './Chat.scss';
-import FriendsList from '../Video/FriendsList.jsx';
-import VideoChat from '../Video/VideoChat.jsx';
 
 const Chat = props => {
 	const {
-		socket, yourID, users, partnerVideo,
+		socket, yourID, users, peerID,
 	} = props;
 	const [messageList, setMessageList] = useState([]);
 	const [newMessage, setNewMessage] = useState('');
-	const [peerID, setPeerID] = useState();
 	const [renderCount, setRenderCount] = useState(0);
 
 	const addToMessageList = (newAuthor, newTime, aNewMessage) => {
@@ -67,30 +64,20 @@ const Chat = props => {
 	}, [socket, yourID, users]);
 	return (
 		<>
-			<div className={styles.container}>
+			<div className={styles.chatContainer}>
+				<div className={styles.chatBox}>
 
-				<div className={styles.friendsList}>
-					<FriendsList users={users} yourID={yourID} messagePeer={setPeerID} callPeer={() => {}} />
+					<div className={styles.messageList}>{messages}</div>
+					<form onSubmit={handleSubmit} className={styles.form}>
+						<input
+							type="text"
+							className={styles.textBox}
+							placeholder="Message"
+							value={newMessage}
+							onChange={event => setNewMessage(event.target.value)}
+						/>
+					</form>
 				</div>
-
-				<div className={styles.chatContainer}>
-					<div className={styles.chatBox}>
-
-						<div className={styles.messageList}>{messages}</div>
-						<form onSubmit={handleSubmit} className={styles.form}>
-							<input
-								type="text"
-								className={styles.textBox}
-								placeholder="Message"
-								value={newMessage}
-								onChange={event => setNewMessage(event.target.value)}
-							/>
-						</form>
-					</div>
-				</div>
-
-				<VideoChat socket={socket} yourID={yourID} partnerVideo={partnerVideo} />
-
 			</div>
 		</>
 	);
